Rework the META file to make indexable by Pinto
Pinto has trouble indexing this package because it doesn't understand
the use of the META.yml in the provides clause of the META info.

The spec also allows referring to the file that generates the package
in that statement, but pinto (apparently following PAUSE's lead)
doesn't consider ReadKey_pm.PL to be a "simile" for that package, the
rule seems to be that ".pm.PL" is the acceptable form.

This commit

- renames ReadKey_pm.PL -> ReadKey.pm.PL;
- touches up the Makefile.PL and MANIFEST; and
- changes the META.yml file to refer to ReadKey.pm.PL

See #29 and thaljef/Pinto/#241.